grader/requirements.txt
# Core Deep Learning
tensorflow>=2.13.0
numpy>=1.24.0

# Configuration & Utilities
pyyaml>=6.0
argparse

# Experiment Tracking
wandb>=0.18.5

# Image Processing
pillow>=10.0.0

# Progress bars & visualization
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Metrics & Evaluation
scikit-learn>=1.3.0

# For reproducibility
python-dotenv>=1.0.0
